Add packet per second (pps) metric
This patch adds pps (packets per second) metric. pps is the ratio of
number of total packets seen to elapsed time.

It is not possible calculating the number of total packets in a pcap
file with a simple iteration. Moreover, calculating this value adds
overhead in the function under testing. The suggested solution is adding
a new value in .ts files recording the value per pcap file. This value
goes under the key 'pcap_pkts'.

* bpf_pflua_ts.lua. calculates pps
* ts/tests/0001_plan.ts. pcap_pkts added