Remove function 'normalize' in MAC address parsing
This function was used for the case that there was only one octet in a
MAC address fragment. Example:

f:f:f:3:3:3

This kind of addresses are not valid MAC addresses but libpcap accepts
them. When there's an address like this, what is missing is the leading
zero, not the trailing zero. The 'normalize' function considered
that the octet missing was the trailing zero, and wrongly shifted the
octet 4 bits.