Clear cache
Currently the values of the cache are not being cleared despite
assigning it an empty map. The reason is that variable 'cfkey_cache' is
assigned an upvalue not a reference to the cache.

Hoisting variable 'cfkey_cache' out of memoize makes possible to reference
the cache and effectively clear it.