Use Rq registers again for most things
This effectively reverts commit 0a4cd5aa8cefcf34a2dd. The problem
is that the IR for pflang allows 64-bit constants in places that
wouldn't be allowed in surface expressions. Since random testing
will generate these IRs, we need to be able to handle them.

This also means stack slots need to be 8 bytes wide since we
may store intermediate results there that need to be 64-bit.